Created a library of MIM functions and an example script
